+++ app/admin/components/vpEditCertifiedProductCtrl.js
@@ -31,22 +31,18 @@
                 vm.workType = workType;
                 vm.showFormErrors = false;
                 vm.message = '';
+                if (vm.cp.chplProductNumber.length > 12) {
+                    var idFields = vm.cp.chplProductNumber.split('.');
+                    vm.idFields = {
+                        prefix: idFields[0] + '.' + idFields[1] + '.' + idFields[2] + '.' + idFields[3],
+                        prod: idFields[4],
+                        ver: idFields[5],
+                        ics: idFields[6],
+                        suffix: idFields[7] + '.' + idFields[8]
+                    };
+                }
 
                 vm.handlers = [];
-
-                if (vm.cp.certificationEdition.name === '2015') {
-                    var idFields = vm.cp.chplProductNumber.split('.');
-                    vm.cp.chplId = {
-                        acb: idFields[0],
-                        atl: idFields[1],
-                        dev: idFields[2],
-                        prod: idFields[3],
-                        ver: idFields[4],
-                        ics: idFields[5],
-                        adds: idFields[6],
-                        date: idFields[7]
-                    }
-                }
                 vm.attachModel();
             }
 
@@ -96,21 +92,37 @@
             }
 
             function save () {
+                if (vm.cp.chplProductNumber.length > 12) {
+                    vm.cp.chplProductNumber =
+                        vm.idFields.prefix + '.' +
+                        vm.idFields.prod + '.' +
+                        vm.idFields.ver + '.' +
+                        vm.idFields.ics + '.' +
+                        vm.idFields.suffix;
+                }
+                vm.cp.certificationDate = vm.cp.certDate.getTime();
                 if (vm.workType === 'manage') {
                     commonService.updateCP(vm.cp)
                         .then(function (response) {
                             if (!response.status || response.status === 200) {
                                 $modalInstance.close(response);
                             } else {
-                                vm.message = response;
+                                vm.errors = [response.error];
                             }
                         },function (error) {
-                            vm.message = response;
+                            vm.errors = [];
+                            if (error.data) {
+                                if (error.data.error && error.data.error.length > 0)
+                                    vm.errors.push(error.data.error);
+                                if (error.data.errorMessages && error.data.errorMessages.length > 0)
+                                    vm.errors = vm.errors.concat(error.data.errorMessages);
+                                if (error.data.warningMessage && error.data.warningMessage.length > 0)
+                                    vm.errors = vm.errors.concat(error.data.warningMessage);
+                            }
                         });
                 } else if (vm.workType === 'confirm') {
                     $modalInstance.close(vm.cp);
                 }
-
             }
         }]);
 })();